REF : LottoMaker 추상화작업
